feat: added ui for committee page under locale after migration
