+++ src/app/(frontend)/[locale]/committee/page.tsx
